requirements-enhanced.txt
# Base requirements
ultralytics>=8.0.20
google-cloud-storage>=2.10.0
label-studio-sdk>=0.0.32
pandas>=2.0.0
tqdm>=4.65.0

# Enhanced pipeline
opencv-python>=4.8.0
scikit-image>=0.21.0
pyyaml>=6.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
click>=8.1.0
watchdog>=3.0.0

# Monitoring & logging
tensorboard>=2.13.0
wandb>=0.15.0
colorlog>=6.7.0
matplotlib>=3.7.0

# Testing & validation
pytest>=7.4.0
pytest-cov>=4.1.0
